6-size.c

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-printf("size of a char: %d bytes(s)\n", (unsigned long)sizeof(char));
-printf("size of int: %d byte(s)\n", (unsigned long)sizeof(int));
-printf("size of long int: %d byte(s)\n", (unsigned long)sizeof(long int));
-printf("size of a long long int: %d byte(s)\n", (unsigned long)sizeof(long long int));
-printf("size of a float:%d byte(s)\n", (unsigned long)sizeof(float));
+printf("size of a char: %d bytes(s)\n", sizeof(char));
+printf("size of int: %d byte(s)\n", sizeof(int));
+printf("size of long int: %d byte(s)\n", sizeof(long int));
+printf("size of a long long int: %d byte(s)\n", sizeof(long long int));
+printf("size of a float:%d byte(s)\n", sizeof(float));
 return (0);
 }
